+++ CHANGELOG.md
@@ -6,6 +6,40 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Add tests for `BeginSponsoringFutureReservesOperation`
+- Add `ChangeTrustAsset`
+- Add `ClawbackOperation`
+- Add `ClawbackClaimableBalanceOperation`
+- Add `LiquidityPoolDepositOperation`
+- Add `LiquidityPoolWithdrawOperation`
+- Add `LiquidityPool` to the `LedgerKey` enum
+- Add `SetTrustlineFlagsOperation`
+- Add support for revoking liquidity pool sponsorship to `RevokeSponsorshipOperation`
+- Add `AUTH_CLAWBACK_ENABLED` to `AccountFlags`
+- Add `TRUSTLINE_CLAWBACK_ENABLED` to `TrustLineFlags`
+- Add `TrustlineAsset`
+- Add `InvalidLiquidityPoolIdLength` to `Error`
+- Add `LedgerKey::LiquidityPool`
+- Add `LiquidityPoolId`
+- Add `LiquidityPoolConstantFeeParameters`
+- Add `LiquidityPoolConstantFee`
+- Add `LiquidityPool`
+- Add `InnerOperationResult::ClaimClaimableBalance`
+- Add `InnerOperationResult::Clawback`
+- Add `InnerOperationResult::ClawbackClaimableBalance`
+- Add `InnerOperationResult::SetTrustLineFlags`
+- Add `InnerOperationResult::LiquidityPoolDeposit`
+- Add `InnerOperationResult::LiquidityPoolWithdraw`
+- Add `InnerTransactionResult::FeeBumpInnerSuccess`
+- Add `InnerTransactionResult::FeeBumpInnerFailed`
+
+### Changed
+
+- BREAKING: Use `ChangeTrustAsset` for `ChangeTrust` operation
+- BREAKING: Use `TrustLineAsset` instead of `Asset` in `LedgerKey::Trustline`
+- BREAKING: Update `xdr_generated.rs` from the current stellar X files.
 
 ## [0.5.0] - 2021-01-15
 ### Added